feat: improve getLatestPosts
